:construction: Move specific extensions to different file
