+++ Toggl.Daneel/Toggl.Daneel.csproj
@@ -627,6 +627,7 @@
     <Compile Include="ViewSources\CreateSuggestionGroupedTableViewSource.cs" />
     <Compile Include="Views\TextViewWithPlaceholder.cs" />
     <Compile Include="Binding\TextViewWithPlaceholderTextTargetBinding.cs" />
+    <Compile Include="Extensions\StartTimeEntryExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Toggl.Foundation.MvvmCross\Toggl.Foundation.MvvmCross.csproj">
